Add measurement data (flue organ pipe and bassoon)
Add measured balloon directivity in form of Matlabs .fig format, for flue pipe (flp) and bassoon (bsn) at several discrete frequenencies as indicated in the filename. For bassoon, data from two different fingerings (340 = MIDI 34 => Bb1, f0 = 58Hz; 530 = MIDI 53 => F3, f0 = 175Hz) is given.
The balloons in bsn340_*.fig are shown in the left column of FIG.2 in the manuscript.
The balloons in flp_*.fig are shown in the left column of FIG.4 in the manuscript.
The balloons in bsn530_*.fig are shown in the left column of FIG.6 in the manuscript.