Add multi-source GUI for up to 14 sources
The model provided within PointSourceGUI_bassoon allows the user to modify the source parameters of up to 14 sources. Source parameters for corresponding to the case of 14 opened tone holes on a bassoon fingered for the note F3 (f0 =175Hz) are provided as numeric tables in *.csv format. These can be loaded as presets in the GUI, and subsequently changed by use of sliders. Frequency dependent radiation balloons are shown and can be exported to file.